requirements.txt
tenseal 
networkx
numpy
pandas
flask
flask_sqlalchemy 
flask_bcrypt 
flask_login
python_dotenv